Moved Pagination to separate component
